Types package simplification for easier type insertion

+++ go/store/types/codec_test.go
@@ -54,7 +54,7 @@ func TestCodecReadFloat(t *testing.T) {
 	test := func(data []byte, exp float64) {
 		r := binaryNomsReader{buff: data}
 		n := r.readFloat(Format_7_18)
-		assert.Equal(t, exp, float64(n))
+		assert.Equal(t, exp, n)
 		assert.Equal(t, len(data), int(r.offset))
 	}
